# src/openapi.yaml
openapi: "3.1.0"
info:
  title: WokAPI
  description: |
    WokAPI is the platform API for the WokSpec ecosystem — providing authentication,
    developer API tokens, billing, and AI routing.

    ## Authentication

    WokAPI supports two authentication methods:

    ### Session auth (cookie)
    Used by the WokSpec dashboard and web apps. Authenticate via GitHub or Google OAuth
    at `/v1/auth/github` or `/v1/auth/google`. The session cookie `wokspec_session` is
    set automatically.

    ### API key auth (Bearer token)
    For developer and programmatic access. Generate keys at `dashboard.wokspec.org/tokens`.

    ```
    Authorization: Bearer wok_live_a1b2c3d4...
    ```

    Keys are prefixed `wok_live_` (production) or `wok_test_` (sandbox).

    ## Rate limits

    All responses include rate limit headers:
    - `X-RateLimit-Limit` — requests allowed per minute
    - `X-RateLimit-Remaining` — requests remaining in current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets

    | Plan | Requests/month | Requests/minute |
    |------|---------------|-----------------|
    | Free | 1,000 | 10 |
    | Pro | 100,000 | 100 |
    | Enterprise | Unlimited | 1,000 |

  version: "1.0.0"
  contact:
    name: WokSpec
    url: https://wokspec.org
    email: hello@wokspec.org
servers:
  - url: https://api.wokspec.org
    description: Production
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: wok_live_xxx
    SessionAuth:
      type: apiKey
      in: cookie
      name: wokspec_session
...
paths:
...
  /v1/auth/github:
    get:
      summary: Initiate GitHub OAuth
      tags: [auth]
      parameters:
        - name: redirect_to
          in: query
          schema:
            type: string
            format: uri
      responses:
        "302":
          description: Redirect to GitHub OAuth
  /v1/auth/me:
    get:
      summary: Get current user
      tags: [auth]
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Current authenticated user
        "401":
          description: Not authenticated
...